+++ src/index.ts
@@ -1,5 +1,4 @@
 export * from './batch'
 export { parameterizedStages } from './behavior'
-export { periodically } from './behavior-presets'
+export { periodically, stochasticPiecewise } from './behavior/presets'
 export * from './local'
-export * from './stochastic-piecewise'
